updete dotfiles

+++ fish/functions/ai_yt.fish
@@ -1,32 +1,32 @@
-function ai_yt
+function ai_yt_de
     set yt_arg $argv[1]
     set fabric_arg $argv[2]
 
-    # Check if yt_arg is provided, if not, prompt with gum
     if test -z "$yt_arg"
         set yt_arg (gum input --placeholder "Enter yt argument:")
     end
 
-    # Execute yt --transcript and store the result
     set yt_result (yt --transcript "$yt_arg")
 
-    # Check if fabric_arg is provided, if not, prompt with gum and list selections
-    if test -z "$fabric_arg"
-        set fabric_arg (fabric --listpatterns | sed '1,2d' | tr -d ' ' | tr -d '\t' | gum filter --placeholder "Select fabric argument:")
+    if test -z "$yt_result"
+        echo "Fehler: Das Ergebnis von yt ist leer. Überprüfe das yt Argument."
+        return
     end
 
+    if test -z "$fabric_arg"
+        set fabric_arg (fabric --listpatterns | sed '1,2d' | tr -d ' ' | gum filter --placeholder "Select fabric argument:")
+    end
 
-    # Create a temporary file
+    set processing_method (gum choose "r_translate" "tana_paste" --placeholder "Select processing method:")
+
     set temp_file (mktemp)
 
-    # Execute fabric -p with the selected argument and append the yt_result, then save to temp_file
-    fabric -p "$fabric_arg" "$yt_result" | fabric -p r_translate > $temp_file
+    if not fabric -p "$fabric_arg" "$yt_result" | fabric -p $processing_method > $temp_file
+        echo "Fehler: fabric Befehl fehlgeschlagen."
+        return
+    end
 
-        cat $temp_file | glow
-
-        # Copy the contents of the processed file to clipboard
-        cat $temp_file | pbcopy
-
-        # Remove the temporary files
-        rm $temp_file
+    cat $temp_file | glow
+    cat $temp_file | pbcopy
+    rm $temp_file
 end
